Fast version of part2
Instead of all multiples of one of the busses, check multiples of the
timestamp where the already-handled busses 'meet'. So for
7,13,x,x,59,x,31,19 the first timestamp is 77, since 78 mod 13 is 0.
That means when finding a timestamp where 7, 13 and 59 line up we can go
with multiples of 77. In turn for 31 lining up in addition it's possible
to go with multiples of 350. The basic idea is that the line-up of n
busses repeats until the n+1th bus also matches up.

This reduces the number of iterations greatly, since the 'stepsize' is
quickly increasing.